+++ src/chain/mod.rs
@@ -17,6 +17,7 @@ use digest::Digest;
 use futures::StreamExt as _;
 use fvm_ipld_blockstore::Blockstore;
 use fvm_ipld_encoding::DAG_CBOR;
+use itertools::Itertools as _;
 use multihash_derive::MultihashDigest as _;
 use nunny::Vec as NonEmpty;
 use serde::{Deserialize, Serialize};
@@ -174,5 +175,25 @@ impl FilecoinSnapshotMetadata {
             head_tipset_key,
             f3_data,
         }
+    }
+}
+
+impl std::fmt::Display for FilecoinSnapshotMetadata {
+    fn fmt(&self, f: &mut std::fmt::Formatter) -> std::fmt::Result {
+        writeln!(f, "Snapshot version: {}", self.version)?;
+        let head_tipset_key_string = self
+            .head_tipset_key
+            .iter()
+            .map(Cid::to_string)
+            .join("\n                  ");
+        writeln!(f, "Head Tipset:      {head_tipset_key_string}")?;
+        write!(
+            f,
+            "F3 data:          {}",
+            self.f3_data
+                .map(|c| c.to_string())
+                .unwrap_or_else(|| "not found".into())
+        )?;
+        Ok(())
     }
 }